myblog/requirements.txt
django==3.0.3
django-debug-toolbar==2.2
mysqlclient==2.0.1
celery==4.4.2
django-celery==3.3.1
lxml==4.5.0
itsdangerous==1.1.0
pymysql==0.9.3
channels==2.4.0
celery==4.4.2
itsdangerous==1.1.0
channels-redis==2.4.2
sqlalchemy==1.3.17
wagtail==2.9.3
Pillow==7.2.0
marshmallow==3.7.1
django-widget-tweaks==1.4.8
Markdown==3.2.2
django-redis==4.7.0